+++ database/012.create-table_user_games.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS user_games (
-  id                  SERIAL          NOT NULL,
-  user_id             INT             NOT NULL,
-  game_id             INT             NOT NULL,
+  id                  BIGSERIAL       NOT NULL,
+  user_id             BIGINT          NOT NULL,
+  game_id             BIGINT          NOT NULL,
   purchase_date       TIMESTAMP,
   total_playtime      INT             DEFAULT 0,
