+++ database/015.create-table_transactions.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS transactions (
-  id                  SERIAL          NOT NULL,
-  user_id             INT             NOT NULL,
-  transaction_date    TIMESTAMP       NOT NULL,
-  total_amount        DECIMAL(14,6)   NOT NULL,
-  status              VARCHAR(20)     NOT NULL,
+  id                  BIGSERIAL             NOT NULL,
+  user_id             BIGINT                NOT NULL,
+  transaction_date    TIMESTAMP             NOT NULL,
+  total_amount        DECIMAL(20,5)         NOT NULL,
+  status              VARCHAR(20)           NOT NULL,
   transaction_type    transaction_type_enum NOT NULL,
   description         VARCHAR(45),
