+++ database/018.create-table_game_discounts.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS game_discounts (
-  game_id             INT             NOT NULL,
-  discount_id         INT             NOT NULL,
+  game_id             BIGINT          NOT NULL,
+  discount_id         BIGINT          NOT NULL,
 
   PRIMARY KEY (game_id, discount_id),
   FOREIGN KEY (game_id) REFERENCES games(id)
